ServletInitializer is configured. Recipe related dtos and repositories are created along with their service implementations.

+++ src/main/java/com/example/recipeapi/category/service/CategoryService.java
@@ -6,7 +6,7 @@ import com.example.recipeapi.category.model.dto.CategoryInsertDto;
 import java.util.List;
 
 public interface CategoryService {
-    List<Category> search(int recipeId);
+    List<Category> search(long recipeId);
 
     void save(CategoryInsertDto insertDto);
 }
